move crawler-parameters-builder to the crawler folder
