+++ packages/cli/src/index.ts
@@ -10,8 +10,8 @@ export class cheerioModule {}
 
 export { AIScanner } from './scanner/ai-scanner';
 export { AICrawler } from './crawler/ai-crawler';
+export { CrawlerParametersBuilder } from './crawler/crawler-parameters-builder';
 export { setupCliContainer } from './setup-cli-container';
 export { AICombinedReportDataConverter } from './converter/ai-data-converter';
-export { CrawlerParametersBuilder } from './crawler-parameters-builder';
 export { ScanArguments } from './scan-arguments';
 export { validateScanArguments } from './validate-scan-arguments';
