+++ packages/cli/src/runner/crawler-command-runner.spec.ts
@@ -8,7 +8,7 @@ import { IMock, It, Mock, Times } from 'typemoq';
 import { ReportDiskWriter } from '../report/report-disk-writer';
 import { ScanArguments } from '../scan-arguments';
 import { ConsolidatedReportGenerator } from '../report/consolidated-report-generator';
-import { CrawlerParametersBuilder } from '../crawler-parameters-builder';
+import { CrawlerParametersBuilder } from '../crawler/crawler-parameters-builder';
 import { AICrawler } from '../crawler/ai-crawler';
 import { CrawlerCommandRunner } from './crawler-command-runner';
